requirements.txt
Django==2.2
psycopg2-binary==2.8.5
python-dotenv==0.14.0
djangorestframework==3.8.2
djangorestframework-simplejwt==3.2.3
pillow==6.2.1
django-extensions==3.0.9
jupyter==1.0.0
notebook==6.1.4
django-rest-auth==0.9.5
django-cors-headers==3.3.0
django-rest-swagger==2.2.0
django-allauth==0.44.0
